+++ changelog.md
@@ -3,6 +3,9 @@
 #### v0.5.1-beta (2023-07-30)
 
 - Omit the 'at least' detail when we witnessed a level up
+- Can now reset the window multiple times with /sofia reset
+- Can now purge roster data multiple times with /sofia delete
+- Fixed almost impercetible rendering issues
 
 #### v0.5.0-beta (2023-07-29)