+++ changelog.md
@@ -1,5 +1,9 @@
 ## SOFIA Changelog
 
+#### v0.6.0-beta (2023-08-25)
+
+- Write in the chat when a guild player levels up every 10 levels
+
 #### v0.5.1-beta (2023-07-30)
 
 - Omit the 'at least' detail when we witnessed a level up